[cw] document, comment, write TODOs, reorganize handlers
