+++ server/core/issue_handlers.go
@@ -1,0 +1,15 @@
+package core
+
+import (
+	"fmt"
+
+	"github.com/gocraft/web"
+)
+
+func (*API) GetIssues(rw web.ResponseWriter, req *web.Request) {
+	fmt.Println("TODO GET ISSUES")
+}
+
+func (*API) GetIssue(rw web.ResponseWriter, req *web.Request) {
+	fmt.Println("TODO GET ISSUE")
+}
