+++ server/core/platform.go
@@ -15,10 +15,10 @@ import (
 // struct for storing in-memory statefulnessnessnesssnes for server
 type Platform struct {
 	*Logger
-	Api         *API
-	Submissions *MagazineAdministrator
-	config      *env.Config
-	db          *sql.DB
+	Api           *API
+	MagazineAdmin *MagazineAdministrator
+	config        *env.Config
+	db            *sql.DB
 }
 
 func NewPlatform() *Platform {
@@ -36,9 +36,9 @@ func NewPlatform() *Platform {
 	// construct API and pass it the db connection handle set within Connect ---^
 	p.Api = NewAPI(p.config, p.db)
 
-	// construct a Submissions system and pass it the db connection handle established above
-	p.Submissions = NewMagazineAdministrator(
-		&p.config.Magazine,
+	// construct a zine admin and pass it the db connection handle established above
+	p.MagazineAdmin = NewMagazineAdministrator(
+		&p.config.MagazineGuidelines,
 		&p.config.ExecContext,
 		p.db,
 	)
@@ -146,7 +146,7 @@ func (p *Platform) Setup() {
 
 func (p *Platform) Start() {
 	// start call for submissions scheduler
-	go p.Submissions.StartScheduler()
+	go p.MagazineAdmin.BeginReleaseCycle()
 
 	// listen on quad-zero route with specified port yo (wait is this garbage?)
 	addr := fmt.Sprintf("0.0.0.0:%s", p.config.Port)
